# jobbing/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Aprende tu mismo API
  description: Application created to provide training courses
  termsOfService: https://www.youtube.com/channel/UCz3p_Kz87WePeRMy6x5CX5w
  contact:
    name: Jalisco con empleo
    url: https://bdalimentos.org/
    email: comunicacionbamx@bdalimentos.org
  license:
    name: Apache 2.0
    url: http://apache.com
  version: 1.0.0
servers:
- url: https://www.jaliscoconempleo.org
  description: Production Server
- url: https://127.0.0.1:4010
  description: Development server (uses test data)
paths:
...
  /user/{uid}:
    get:
      tags:
      - User
      description: Displays a User defined by ID
      operationId: get_user_by_id
      parameters:
      - name: uid
        in: path
        description: Unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"BAD_REQUEST\"}"
                x-content-type: application/json
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Unauthorized\"}"
                x-content-type: application/json
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"Error 404, page not found.\"}"
                x-content-type: application/json
        "500":
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                type: string
                example: "{\"message\":\"unexpected error\"}"
                x-content-type: application/json
      x-openapi-router-controller: jobbing.controllers.user_controller
...
components:
  schemas:
...
    User:
      required:
      - email
      - userId
      - username
      type: object
      properties:
        userId:
          type: integer
          description: User id to index table
          readOnly: true
          example: 123
        uid:
          type: string
          description: Public id
          format: uid
          readOnly: true
          example: k8df23
        username:
          type: string
          description: Unique name for user
          example: nixabell1835
        password:
          type: string
          description: Keyword to access the app
          format: password
          writeOnly: true
        passwordDate:
          type: string
          description: Last password updated date
          format: date
          readOnly: true
        email:
          type: string
          description: e-mail related to this account
          format: email
        imageProfile:
          type: string
          description: An image profile for the user
          nullable: true
          example: "0"
        roleId:
          type: integer
          description: Foreign key for the role type assigned to this user
          readOnly: true
        providerId:
          type: integer
          description: Foreign key to User Profile
      example:
        uid: k8df23
        password: ""
        roleId: 0
        imageProfile: "0"
        passwordDate: 2000-01-23
        userId: 123
        email: ""
        username: nixabell1835
        providerId: 3
...